+++ lib/ohm/sorted.rb
@@ -105,8 +105,12 @@ module Ohm
         model.db
       end
 
-      def each(&block)
-        ids.each { |id| block.call(model.to_proc[id]) }
+      def each
+        if block_given?
+          ids.each { |id| block.call(model.to_proc[id]) }
+        else
+          to_enum(:each)
+        end
       end
 
       def [](id)